Add conditional imports (#36)
* Add conditional imoprts for main.dart

* Add conditional imports for Drift
