+++ lib/main.dart
@@ -1,11 +1,6 @@
-import 'package:purple_starter/src/feature/app/logic/main_runner.dart';
-import 'package:purple_starter/src/feature/app/model/async_app_dependencies.dart';
-import 'package:purple_starter/src/feature/app/purple_starter_app.dart';
+// ignore: unused_import
+import 'package:purple_starter/runner_stub.dart'
+    if (dart.library.io) 'package:purple_starter/runner_io.dart'
+    if (dart.library.html) 'package:purple_starter/runner_web.dart' as runner;
 
-Future<void> main() => MainRunner.run<AsyncAppDependencies>(
-      asyncDependencies: AsyncAppDependencies.obtain,
-      appBuilder: (sentrySubscription, dependencies) => PurpleStarterApp(
-        sentrySubscription: sentrySubscription,
-        sharedPreferences: dependencies.sharedPreferences,
-      ),
-    );
+Future<void> main() => runner.run();
