+++ lib/runner_io.dart
@@ -1,0 +1,3 @@
+import 'package:purple_starter/runner_shared.dart' as runner;
+
+Future<void> run() => runner.run();
