+++ lib/runner_stub.dart
@@ -1,0 +1,3 @@
+import 'package:purple_starter/src/core/error/unknown_host_platform_error.dart';
+
+Future<void> run() => throw UnknownHostPlatformError();
