+++ lib/src/core/database/drift/connection/open_connection_io.dart
@@ -5,19 +5,9 @@ import 'package:drift/native.dart';
 import 'package:path/path.dart' as path;
 import 'package:path_provider/path_provider.dart';
 
-part 'app_database.g.dart';
-
-LazyDatabase _openConnection(String name) => LazyDatabase(() async {
+QueryExecutor openConnection(String name) => LazyDatabase(() async {
       final dbFolder = await getApplicationDocumentsDirectory();
       final file = File(path.join(dbFolder.path, '$name.sqlite'));
 
       return NativeDatabase(file);
     });
-
-@DriftDatabase()
-class AppDatabase extends _$AppDatabase {
-  AppDatabase({required String name}) : super(_openConnection(name));
-
-  @override
-  int get schemaVersion => 1;
-}
