+++ lib/src/core/database/drift/connection/open_connection_stub.dart
@@ -1,0 +1,4 @@
+import 'package:drift/drift.dart';
+import 'package:purple_starter/src/core/error/unknown_host_platform_error.dart';
+
+QueryExecutor openConnection(String name) => throw UnknownHostPlatformError();
