+++ lib/src/core/database/drift/connection/open_connection_web.dart
@@ -1,0 +1,4 @@
+import 'package:drift/drift.dart';
+import 'package:drift/web.dart';
+
+QueryExecutor openConnection(String name) => WebDatabase(name);
