+++ lib/src/core/error/unknown_host_platform_error.dart
@@ -1,0 +1,3 @@
+class UnknownHostPlatformError extends UnsupportedError {
+  UnknownHostPlatformError() : super('Unknown host platform.');
+}
