+++ lib/src/core/extension/src/build_context.dart
@@ -2,7 +2,7 @@ import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 // ignore: depend_on_referenced_packages
 import 'package:flutter_gen/gen_l10n/app_localizations.g.dart';
-import 'package:purple_starter/src/core/database/app_database.dart';
+import 'package:purple_starter/src/core/database/drift/app_database.dart';
 import 'package:purple_starter/src/core/model/dependencies_storage.dart';
 import 'package:purple_starter/src/core/model/repository_storage.dart';
 import 'package:purple_starter/src/core/widget/dependencies_scope.dart';
