+++ lib/src/feature/settings/database/settings_dao.dart
@@ -1,4 +1,4 @@
-import 'package:purple_starter/src/core/database/shared_preferences_dao.dart';
+import 'package:purple_starter/src/core/database/shared_preferences/shared_preferences_dao.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 abstract class ISettingsDao {
